+++ CHANGELOG.md
@@ -13,8 +13,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/), and this
 - Enabled JSON serialization for frontend-compatible error responses.
 - Added support for logging services like Sentry.
 
-## [1.1.0] - 2024-10-29
+## [1.0.1] - 2024-10-29
 ### Added
 - Option to customize error responses using `formatError`.
 - Improved error handling for unsupported HTTP methods.
 - Expanded documentation in the README.
+
+## [1.1.9] - 2024-10-30
+### Fixed
+- Enhanced the `errorHandler` function to support both API Routes and App Router in Next.js, allowing for flexible error handling across different route types.
+- Adjusted error response handling to ensure appropriate responses are returned based on the presence of `res` in the request.